Dataset updates: sharding, dummy data, data dict
A few things ended up entangled in this commit and hard to pick apart.

I added support for dataset sharding across workers.

Dummy dataset implementation was finished and works.

The get_datasets functions now return dictionaries of the datasets and
some finalized data config which previously wasn't available and/or
needed to be re-computed.