update options and logic for staging again
This cleans up the internal logic a bit. Only one check on the
divisibility of the samples is needed, and the options for sharding and
local-staging are once again separate.

The command line argument --shard-type replaced with --staged-files. I'm
assuming that it's good enough to set the shard flag via the config
files, as there are very limited cases in which you wouldn't shard the
data.