Moved to fully custom LR schedule and new config
Moved away from Horovod's LR callbacks completely.

Now just using the Keras LR scheduler callback with a function that
handles the linear warmup and the decay steps.

The configuration now has a lr_schedule section for setting everything
related to LR scheduling. One can still set the LR on the optimizer if
the schedule is not desired.

LR scaling is now specified not in terms of ranks but in terms of batch
size, which I think makes a lot more sense. So, in the config, you
specify a base_lr and a base_batch_size, in addition to the scaling
function (linear or sqrt). The ratio of global batch size to base batch
size gives the scale factor, replacing n_ranks from before.